Updates from Zax_Branch
Added:
movingAI instances to resources

Changes:
Moved methods from InstanceBuilder_BGU as static methods in interface I_InstanceBuilder

Implements:
Started implementing InstanceBuilder_MovingAI (Only maps for now)

+++ src/test/java/Instances/InstanceBuilder_BGUTest.java
@@ -21,7 +21,7 @@ public class InstanceBuilder_BGUTest {
     private InstanceBuilder_BGU instanceBuilderBgu;
     private InstanceManager.InstancePath instancePath;
     private InstanceProperties instanceProperties = new InstanceProperties(
-                                                        new int[]{16,16},0,7,"-"
+                                                        new int[]{16,16},(float)0,7,"-"
                                                     );
 
     private final String path = IO_Manager.buildPath(
